requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI & ML
requests==2.31.0
httpx==0.25.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Monitoring & Metrics
prometheus-client==0.19.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0

# Utilities
pydantic==2.5.2
pydantic-settings==2.0.3
python-dateutil==2.8.2
schedule==1.2.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0

# Added from the code block
aiohttp==3.9.1
